WINGs: Do not call exit()
WMCreateFont() was calling exit() if it could not create the
font, and was trying too hard not to return NULL.

Just return NULL if the font could not be created instead of exit()ing
and let callers decide what to do upon failure.

Thanks to Christian <chris@computersalat.de> for reporting this.