WindowMaker: New get_wwindow_image_from_x11 function
A new function is created to get the image from the X11 window.

The function updateIconImage is splitted in two blocks, one is moved
to get_wwindow_image_from_x11 with the X11 code, and other is used
to update the Application and Window icons (kept in the same function).