Also relaunch from appicon with Ctrl + DblClick.
Zoltan Balaton points out that Control + Doubleclick on docked app will
launch a new instance.  For consistency with that behaviour we now allow
Control + Doubleclick on an undocked appicon to invoke the new
relaunching functionality.

We also now restrict doubleclick handling to the left mouse button in
order to avoid relaunching the application twice when the middle button
is used.