Fixed wrong structure declaration
The declaration of the structure actually also created an unused variable. This
variable was not used anywhere, and lead to symbol defined in multiple objects.
These symbol are silently merged by GCC, thus no problem was reported.

The issue was raised by Yves de Champlain when trying to compile with LLVM/clang
which is a bit stricter there.